hf: sum interest masses in float64 to match the full-vocab path exactly
The gathered probabilities were summed in float32 while the legacy
full-vocab path accumulates Python floats, so tokens backed by several
vocabulary ids could drift a few 1e-8 from the distribution they claim
parity with -- enough to flake the 1e-9 parity test.

+++ s3e/vlm/huggingface.py
@@ -247,7 +247,9 @@ class HuggingFaceVLM(VLMBackend):
             spans.append((token, len(all_ids), len(all_ids) + len(ids)))
             all_ids.extend(ids)
 
-        selected = probs[:, all_ids].cpu() if all_ids else None
+        # float64 so summing duplicate ids matches the full-vocab path, which
+        # accumulates Python floats.
+        selected = probs[:, all_ids].double().cpu() if all_ids else None
         argmax_ids = probs.argmax(dim=-1).cpu().tolist()
         interest_id_set = set(all_ids)
